lambda-ingestion-irceline-api/requirements.txt
python-dotenv
pymysql
requests